requirements-test.txt
# Lint requirements
--requirement requirements-lint.txt

# Build requirements
cython
cmake>=3.26
setuptools
scikit-build-core
ninja

# Runtime requirements
--requirement requirements.txt

# Test requirements
cffi
cpplint
cython
docutils
dtlib
numpy>=1.23.5
pytest>=6.2.4
pytest_xdist>=2.2.1
packaging>=21.0
PyYAML
tqdm>=4.62.3
typing_extensions>=4.10.0
requests
cloudpickle
ml_dtypes
psutil
torch
tabulate
wheel
setuptools
einops
scipy
tornado
ninja==1.10.0
cuda-python>=12.0.0
pytest-xdist>=2.2.1
pytest-durations
pytest-timeout
pyyaml
